Separate live send pipeline context contracts
