+++ src/PlateauResoniteLink/Targets/Resonite/LiveSendFinalizationContext.cs
@@ -1,0 +1,11 @@
+using System;
+
+using PlateauResoniteLink.Transport.ResoniteLink;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record LiveSendFinalizationContext(
+    Uri Endpoint,
+    LiveSendEnqueueContext EnqueueContext,
+    ResoniteLinkSendDiagnostics Diagnostics,
+    Action<string>? ProgressReporter);
