+++ src/PlateauResoniteLink/Targets/Resonite/LiveSendSetupInitialization.cs
@@ -1,0 +1,7 @@
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record LiveSendSetupInitialization(
+    ResoniteSceneSetupState SetupState,
+    LiveSendProgressSink Progress,
+    CommonMaterialAssetCache Materials,
+    ResoniteSharedSlotIndex Placement);
