+++ src/PlateauResoniteLink/Targets/Resonite/LiveSendWorkerLaunchRequest.cs
@@ -1,0 +1,7 @@
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record LiveSendWorkerLaunchRequest(
+    LiveSendRunState State,
+    LiveSendQueuePlan QueuePlan,
+    ResoniteImportBudgetProfile ResourceBudget,
+    int ConnectionCount);
